+++ doc/source/changelog.rst
@@ -14,6 +14,7 @@ This is release version of ``0.9.0`` of Nuclei SDK, which is still under develop
   - Enable __LD/__SD macro when Zilsd extension present for rv32 in ``core_feature_base.h``
   - Add comments for updating ECLIC threshold MTH setting recommendations in ``core_feature_eclic.h``
   - Fix **HDBG** bit position in SysTimer **MTIMECTL** from 4 to 3 in ``core_feature_timer.h``
+  - Add **MTIME_SRC** bit position and update SRW control handling in ``core_feature_timer.h``
 
 * Application